New dir struc and model
During the process of adding the new model nsmod_switching_torque_
with_Euler it was realised that the way the project was structred
didn't make sense.

We now have a dir nsmod that contains all the "physics" a dir "src"
containing the source files. The whole repo name will be changed next
So that we can include results in this repo. Once added to the python
path this will mean imports can be made in the normal manner.