Updates to the Lyne analytic calculation
It appears this work may be flawed: while it is instructuve to derive
the phase residual in such an intuitive manor, the least-sqaures method
does not work. Probably becuase it lacks the idea of "fitting over
several cycles". I can't seem to get the analytic and numeric results to
exactly agree so it is only really useful for the mismatch calculation
as a sort of "approximation".